Move beacon healing to its own module
