+++ src/Parser/HolyPaladin/Modules/PaladinCore/BeaconHealing.js
@@ -1,0 +1,78 @@
+import React from 'react';
+
+import SPELLS from 'common/SPELLS';
+import SpellIcon from 'common/SpellIcon';
+import { formatPercentage } from 'common/format';
+
+import Module from 'Parser/Core/Module';
+
+import StatisticBox, { STATISTIC_ORDER } from 'Main/StatisticBox';
+
+import CPM_ABILITIES  from '../../CPM_ABILITIES';
+
+class BeaconHealing extends Module {
+  getTotalHealsOnBeaconPercentage(parser) {
+    const abilityTracker = parser.modules.abilityTracker;
+    const getCastCount = spellId => abilityTracker.getAbility(spellId);
+
+    const selectedCombatant = parser.selectedCombatant;
+    if (!selectedCombatant) {
+      return null;
+    }
+
+    let casts = 0;
+    let castsOnBeacon = 0;
+
+    CPM_ABILITIES
+      .filter(ability => ability.isActive === undefined || ability.isActive(selectedCombatant))
+      .forEach((ability) => {
+        const castCount = getCastCount(ability.spell.id);
+        casts += (ability.getCasts ? ability.getCasts(castCount) : castCount.casts) || 0;
+        castsOnBeacon += castCount.healingBeaconHits || 0;
+      });
+
+    return castsOnBeacon / casts;
+  }
+
+  suggestions(when) {
+    const totalHealsOnBeaconPercentage = this.getTotalHealsOnBeaconPercentage(this.owner);
+
+    when(totalHealsOnBeaconPercentage).isGreaterThan(0.2)
+      .addSuggestion((suggest, actual, recommended) => {
+        return suggest('You cast a lot of direct heals on beacon targets. Direct healing beacon targets is inefficient. Try to only cast on beacon targets when they would otherwise die.')
+          .icon('ability_paladin_beaconoflight')
+          .actual(`${formatPercentage(totalHealsOnBeaconPercentage)}% of all your healing spell casts were on a beacon target`)
+          .recommended(`<${formatPercentage(recommended)}% is recommended`)
+          .regular(recommended + 0.05).major(recommended + 0.15);
+      });
+  }
+  statistic() {
+    const abilityTracker = this.owner.modules.abilityTracker;
+    const getAbility = spellId => abilityTracker.getAbility(spellId);
+
+    const flashOfLight = getAbility(SPELLS.FLASH_OF_LIGHT.id);
+    const holyLight = getAbility(SPELLS.HOLY_LIGHT.id);
+
+    const flashOfLightHeals = flashOfLight.casts || 0;
+    const holyLightHeals = holyLight.casts || 0;
+    const totalFolsAndHls = flashOfLightHeals + holyLightHeals;
+
+    const beaconFlashOfLights = flashOfLight.healingBeaconHits || 0;
+    const beaconHolyLights = holyLight.healingBeaconHits || 0;
+    const totalFolsAndHlsOnBeacon = beaconFlashOfLights + beaconHolyLights;
+    const totalHealsOnBeaconPercentage = this.getTotalHealsOnBeaconPercentage(this.owner);
+
+    return (
+      <StatisticBox
+        icon={<SpellIcon id={this.owner.selectedCombatant.lv100Talent} />}
+        value={`${formatPercentage(totalFolsAndHlsOnBeacon / totalFolsAndHls)} %`}
+        label="FoL/HL cast on beacon"
+        tooltip={`The amount of Flash of Lights and Holy Lights cast on beacon targets. You cast ${beaconFlashOfLights} Flash of Lights and ${beaconHolyLights} Holy Lights on beacon targets.<br /><br />
+            Your total heals on beacons was <b>${formatPercentage(totalHealsOnBeaconPercentage)}%</b> (this includes spell other than FoL and HL).`}
+      />
+    );
+  }
+  statisticOrder = STATISTIC_ORDER.CORE(50);
+}
+
+export default BeaconHealing;
